Move Statsd into GitHub module
To avoid collision with other top level Statsd classes.

Also gets the specs passing again and adds a gemfile to make it easier to work on.
