+++ spec/helper.rb
@@ -3,7 +3,7 @@ require 'minitest/autorun'
 
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
-require 'statsd'
+require 'github/statsd'
 require 'logger'
 
 class FakeUDPSocket
