+++ spec/statsd_spec.rb
@@ -1,39 +1,39 @@
-require 'helper'
+require_relative './helper'
 
-describe Statsd do
+describe GitHub::Statsd do
   before do
-    @statsd = Statsd.new('localhost', 1234)
+    @statsd = GitHub::Statsd.new(FakeUDPSocket)
+    @statsd.add_shard
     class << @statsd
       public :sampled # we need to test this
-      attr_reader :host, :port # we also need to test this
-      def socket; @socket ||= FakeUDPSocket.new end
     end
   end
 
-  after { @statsd.socket.clear }
+  after { @statsd.shards.first.clear }
 
   describe "#initialize" do
-    it "should set the host and port" do
-      @statsd.host.must_equal 'localhost'
-      @statsd.port.must_equal 1234
+    it "should default client class to UDPClient" do
+      statsd = GitHub::Statsd.new
+      statsd.client_class.must_equal GitHub::Statsd::UDPClient
     end
 
-    it "should default the port to 8125" do
-      Statsd.new('localhost').instance_variable_get('@port').must_equal 8125
+    it "should allow changing client class" do
+      statsd = GitHub::Statsd.new(FakeUDPSocket)
+      statsd.client_class.must_equal FakeUDPSocket
     end
   end
 
   describe "#increment" do
     it "should format the message according to the statsd spec" do
       @statsd.increment('foobar')
-      @statsd.socket.recv.must_equal ['foobar:1|c']
+      @statsd.shards.first.recv.must_equal ['foobar:1|c']
     end
 
     describe "with a sample rate" do
       before { class << @statsd; def rand; 0; end; end } # ensure delivery
       it "should format the message according to the statsd spec" do
         @statsd.increment('foobar', 0.5)
-        @statsd.socket.recv.must_equal ['foobar:1|c|@0.5']
+        @statsd.shards.first.recv.must_equal ['foobar:1|c|@0.5']
       end
     end
   end
@@ -41,14 +41,14 @@ describe Statsd do
   describe "#decrement" do
     it "should format the message according to the statsd spec" do
       @statsd.decrement('foobar')
-      @statsd.socket.recv.must_equal ['foobar:-1|c']
+      @statsd.shards.first.recv.must_equal ['foobar:-1|c']
     end
 
     describe "with a sample rate" do
       before { class << @statsd; def rand; 0; end; end } # ensure delivery
       it "should format the message according to the statsd spec" do
         @statsd.decrement('foobar', 0.5)
-        @statsd.socket.recv.must_equal ['foobar:-1|c|@0.5']
+        @statsd.shards.first.recv.must_equal ['foobar:-1|c|@0.5']
       end
     end
   end
@@ -56,14 +56,14 @@ describe Statsd do
   describe "#timing" do
     it "should format the message according to the statsd spec" do
       @statsd.timing('foobar', 500)
-      @statsd.socket.recv.must_equal ['foobar:500|ms']
+      @statsd.shards.first.recv.must_equal ['foobar:500|ms']
     end
 
     describe "with a sample rate" do
       before { class << @statsd; def rand; 0; end; end } # ensure delivery
       it "should format the message according to the statsd spec" do
         @statsd.timing('foobar', 500, 0.5)
-        @statsd.socket.recv.must_equal ['foobar:500|ms|@0.5']
+        @statsd.shards.first.recv.must_equal ['foobar:500|ms|@0.5']
       end
     end
   end
@@ -71,7 +71,7 @@ describe Statsd do
   describe "#time" do
     it "should format the message according to the statsd spec" do
       @statsd.time('foobar') { sleep(0.001); 'test' }
-      data = @statsd.socket.recv
+      data = @statsd.shards.first.recv
       key, value, unit = data.first.split(/[:|]/)
       key.must_equal "foobar"
       value.must_match /^\d\.\d{3}$/
@@ -88,7 +88,7 @@ describe Statsd do
 
       it "should format the message according to the statsd spec" do
         result = @statsd.time('foobar', 0.5) { sleep(0.001); 'test' }
-        data = @statsd.socket.recv
+        data = @statsd.shards.first.recv
         key, value, unit, frequency = data.first.split(/[:|]/)
         key.must_equal "foobar"
         value.must_match /^\d\.\d{3}$/
@@ -132,40 +132,18 @@ describe Statsd do
 
     it "should add namespace to increment" do
       @statsd.increment('foobar')
-      @statsd.socket.recv.must_equal ['service.foobar:1|c']
+      @statsd.shards.first.recv.must_equal ['service.foobar:1|c']
     end
 
     it "should add namespace to decrement" do
       @statsd.decrement('foobar')
-      @statsd.socket.recv.must_equal ['service.foobar:-1|c']
+      @statsd.shards.first.recv.must_equal ['service.foobar:-1|c']
     end
 
     it "should add namespace to timing" do
       @statsd.timing('foobar', 500)
-      @statsd.socket.recv.must_equal ['service.foobar:500|ms']
+      @statsd.shards.first.recv.must_equal ['service.foobar:500|ms']
     end
-  end
-
-  describe "with logging" do
-    require 'stringio'
-    before { Statsd.logger = Logger.new(@log = StringIO.new)}
-
-    it "should write to the log in debug" do
-      Statsd.logger.level = Logger::DEBUG
-
-      @statsd.increment('foobar')
-
-      @log.string.must_match "Statsd: foobar:1|c"
-    end
-
-    it "should not write to the log unless debug" do
-      Statsd.logger.level = Logger::INFO
-
-      @statsd.increment('foobar')
-
-      @log.string.must_be_empty
-    end
-
   end
 
   describe "stat names" do
@@ -175,29 +153,29 @@ describe Statsd do
     end
 
     it "should replace ruby constant delimeter with graphite package name" do
-      class Statsd::SomeClass; end
-      @statsd.increment(Statsd::SomeClass, 1)
+      class GitHub::Statsd::SomeClass; end
+      @statsd.increment(GitHub::Statsd::SomeClass, 1)
 
-      @statsd.socket.recv.must_equal ['Statsd.SomeClass:1|c']
+      @statsd.shards.first.recv.must_equal ['GitHub.Statsd.SomeClass:1|c']
     end
 
     it "should replace statsd reserved chars in the stat name" do
       @statsd.increment('ray@hostname.blah|blah.blah:blah', 1)
-      @statsd.socket.recv.must_equal ['ray_hostname.blah_blah.blah_blah:1|c']
+      @statsd.shards.first.recv.must_equal ['ray_hostname.blah_blah.blah_blah:1|c']
     end
 
   end
 
 end
 
-describe Statsd do
+describe GitHub::Statsd do
   describe "with a real UDP socket" do
     it "should actually send stuff over the socket" do
       socket = UDPSocket.new
       host, port = 'localhost', 12345
       socket.bind(host, port)
 
-      statsd = Statsd.new(host, port)
+      statsd = GitHub::Statsd.new(host, port)
       statsd.increment('foobar')
       message = socket.recvfrom(16).first
       message.must_equal 'foobar:1|c'
